requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.5
django-cors-middleware==1.5.0
django-filter==2.3.0
django-import-export==2.1.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
PyJWT==1.7.1
PyMySQL==0.10.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
